refactor(vendingMachine): 구매 가능 여부 검증 로직 수정 및 디렉토리 위치 변경
